refactor: Change loop index type in InteractionTestLayer for better type safety
- Updated the loop index in `InteractionTestLayer` from `int` to `size_t` to ensure compatibility with the size of the `controllers_` vector, enhancing type safety and preventing potential issues with negative indices.

+++ samples/glfw_opengl/03_test_interaction/demo_test_interaction.cpp
@@ -67,7 +67,7 @@ class InteractionTestLayer : public vne::testbed::ILayer {
     InteractionTestLayer()
         : vne::testbed::ILayer("InteractionTestLayer") {
         controllers_.resize(BaseSceneLayer::kMaxViewports);
-        for (int i = 0; i < BaseSceneLayer::kMaxViewports; ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(BaseSceneLayer::kMaxViewports); ++i) {
             controllers_[i] = std::make_unique<vne::interaction::CameraSystemController>(
                 vne::interaction::CameraManipulatorType::eOrbitArcball);
         }
